test: apadt tests add user
Moved tests add user from EditTravelSettings to ParticipantsListScreen.
