+++ app/src/androidTest/java/com/github/se/travelpouch/ui/travel/EditTravelScreenTest.kt
@@ -14,7 +14,6 @@ import com.github.se.travelpouch.model.location.LocationRepository
 import com.github.se.travelpouch.model.location.LocationViewModel
 import com.github.se.travelpouch.model.notifications.NotificationRepository
 import com.github.se.travelpouch.model.notifications.NotificationViewModel
-import com.github.se.travelpouch.model.profile.Profile
 import com.github.se.travelpouch.model.profile.ProfileModelView
 import com.github.se.travelpouch.model.profile.ProfileRepository
 import com.github.se.travelpouch.model.travels.ListTravelViewModel
@@ -23,21 +22,17 @@ import com.github.se.travelpouch.model.travels.Participant
 import com.github.se.travelpouch.model.travels.Role
 import com.github.se.travelpouch.model.travels.TravelContainer
 import com.github.se.travelpouch.model.travels.TravelRepository
-import com.github.se.travelpouch.model.travels.fsUid
 import com.github.se.travelpouch.ui.navigation.NavigationActions
 import com.google.firebase.Timestamp
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.mockito.Mockito.doAnswer
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.verify
 import org.mockito.Mockito.`when`
 import org.mockito.kotlin.any
-import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.atLeastOnce
 import org.mockito.kotlin.doNothing
-import org.mockito.kotlin.never
 
 class EditTravelSettingsScreenTest {
   // Helper function to input text into a text field
@@ -95,21 +90,26 @@ class EditTravelSettingsScreenTest {
     profileRepository = mock(ProfileRepository::class.java)
     profileModelView = ProfileModelView(profileRepository)
     locationRepository = mock(LocationRepository::class.java)
-      locationViewModel = LocationViewModel(locationRepository)
+    locationViewModel = LocationViewModel(locationRepository)
   }
 
   @Test
   fun checkNoSelectedTravel() {
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
 
     composeTestRule.onNodeWithTag("editScreen").assertIsDisplayed()
     composeTestRule.onNodeWithTag("editTravelText").assertIsDisplayed()
     composeTestRule.onNodeWithTag("editTravelText").assertTextEquals("Edit Travel")
     composeTestRule.onNodeWithTag("goBackButton").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserFab").assertIsDisplayed()
+    composeTestRule.onNodeWithTag("plusButton").assertIsDisplayed().performClick()
+    composeTestRule.onNodeWithTag("manageParticipantsButton").assertIsDisplayed()
     composeTestRule.onNodeWithTag("importEmailFab").assertIsDisplayed()
     composeTestRule
         .onNodeWithTag("noTravelSelectedText")
@@ -125,7 +125,11 @@ class EditTravelSettingsScreenTest {
         travelContainer) // this causes strange overwrite, shouldn't happen IRL
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
 
     composeTestRule.onNodeWithTag("editScreen").assertIsDisplayed()
@@ -138,14 +142,13 @@ class EditTravelSettingsScreenTest {
     composeTestRule.onNodeWithTag("inputTravelTitle").assertIsDisplayed()
     composeTestRule.onNodeWithTag("inputTravelDescription").assertIsDisplayed()
     composeTestRule.onNodeWithTag("inputTravelLocationName").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("inputTravelLatitude").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("inputTravelLongitude").assertIsDisplayed()
     composeTestRule.onNodeWithTag("inputTravelStartTime").assertIsDisplayed()
     composeTestRule.onNodeWithTag("inputTravelEndTime").assertIsDisplayed()
     composeTestRule.onNodeWithTag("travelSaveButton").performScrollTo().assertIsDisplayed()
     composeTestRule.onNodeWithTag("travelDeleteButton").performScrollTo().assertIsDisplayed()
     composeTestRule.onNodeWithTag("goBackButton").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserFab").assertIsDisplayed()
+    composeTestRule.onNodeWithTag("plusButton").assertIsDisplayed().performClick()
+    composeTestRule.onNodeWithTag("manageParticipantsButton").assertIsDisplayed()
     composeTestRule.onNodeWithTag("importEmailFab").assertIsDisplayed()
   }
 
@@ -155,14 +158,16 @@ class EditTravelSettingsScreenTest {
     listTravelViewModel.selectTravel(travelContainer)
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
 
     inputText("inputTravelTitle", travelContainer.title, "Test Title")
     inputText("inputTravelDescription", travelContainer.description, "Test Description")
     inputText("inputTravelLocationName", travelContainer.location.name, "Test Location")
-    inputText("inputTravelLatitude", travelContainer.location.latitude.toString(), "12.34")
-    inputText("inputTravelLongitude", travelContainer.location.longitude.toString(), "56.78")
     inputText("inputTravelStartTime", "14/02/2009", "14/02/2009")
     inputText("inputTravelEndTime", "16/02/2009", "15/02/2009")
     composeTestRule.onNodeWithTag("travelSaveButton").performScrollTo().assertIsDisplayed()
@@ -174,275 +179,24 @@ class EditTravelSettingsScreenTest {
   }
 
   @Test
-  fun addUserButtonFails() {
-    val travelContainer = createContainer()
-    listTravelViewModel.selectTravel(travelContainer)
-    composeTestRule.setContent {
-      EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
-    }
-    composeTestRule.onNodeWithTag("importEmailFab").performClick()
-    composeTestRule.onNodeWithTag("addUserFab").performClick()
-
-    // perform add user
-    // Check that the dialog is displayed
-    composeTestRule.onNodeWithTag("roleDialogColumn").assertIsDisplayed()
-    // Check that the title text is displayed and correct
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertTextEquals("Add User by Email")
-    // Check that the OutlinedTextField is displayed and has the correct default value
-    composeTestRule.onNodeWithTag("addUserEmailField").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserEmailField").assertTextContains("")
-
-    val randomEmail = "random.email@example.org"
-    inputText("addUserEmailField", "", randomEmail)
-    // Now this is an invalid user that doesn't exist
-    doAnswer { invocation ->
-          val onFailure = invocation.getArgument<(Exception) -> Unit>(2)
-          onFailure(Exception("Unknown API Error"))
-        }
-        .`when`(profileRepository)
-        .getFsUidByEmail(any(), any(), any())
-
-    doAnswer { "abcdefghijklmnopqrst" }.`when`(notificationRepository).getNewUid()
-
-    composeTestRule.onNodeWithTag("addUserButton").performClick()
-    verify(profileRepository).getFsUidByEmail(anyOrNull(), anyOrNull(), anyOrNull())
-    verify(notificationRepository, never()).addNotification(anyOrNull())
-  }
-
-  @Test
-  fun addUserButtonWithUserUid() {
-    val travelContainer = createContainer()
-
-    val profile =
-        Profile(
-            "qwertzuiopasdfghjklyxcvbnm12",
-            "username",
-            "email@gmail.com",
-            null,
-            "name",
-            emptyList())
-
-    `when`(profileRepository.getProfileElements(anyOrNull(), anyOrNull())).then {
-      it.getArgument<(Profile) -> Unit>(0)(profile)
-    }
-    profileModelView.getProfile()
-
-    listTravelViewModel.selectTravel(travelContainer)
-    composeTestRule.setContent {
-      EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
-    }
-    composeTestRule.onNodeWithTag("importEmailFab").performClick()
-    composeTestRule.onNodeWithTag("addUserFab").performClick()
-
-    // perform add user
-    // Check that the dialog is displayed
-    composeTestRule.onNodeWithTag("roleDialogColumn").assertIsDisplayed()
-    // Check that the title text is displayed and correct
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertTextEquals("Add User by Email")
-    // Check that the OutlinedTextField is displayed and has the correct default value
-    composeTestRule.onNodeWithTag("addUserEmailField").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserEmailField").assertTextContains("")
-
-    val randomEmail = "random.email@example.org"
-    inputText("addUserEmailField", "", randomEmail)
-
-    doAnswer { invocation ->
-          val email = invocation.getArgument<String>(0)
-          val onSuccess = invocation.getArgument<(fsUid?) -> Unit>(1)
-          val customUserInfo =
-              Profile(
-                  fsUid = profileModelView.profile.value.fsUid,
-                  name = "Custom User",
-                  userTravelList = listOf("00000000000000000000"),
-                  email = email,
-                  username = "username",
-                  friends = null)
-          // Call the onSuccess callback with the custom UserInfo
-          onSuccess(customUserInfo.fsUid)
-        }
-        .`when`(profileRepository)
-        .getFsUidByEmail(any(), any(), any())
-
-    doAnswer { "abcdefghijklmnopqrst" }.`when`(notificationRepository).getNewUid()
-
-    // Mock the repository.updateTravel method to do nothing
-    doNothing().`when`(travelRepository).updateTravel(any(), any(), any())
-    doAnswer { "sigmasigmasigmasigm2" }.`when`(travelRepository).getNewUid()
-    composeTestRule.onNodeWithTag("addUserButton").performClick()
-
-    verify(profileRepository).getFsUidByEmail(anyOrNull(), anyOrNull(), anyOrNull())
-    verify(notificationRepository, never()).addNotification(anyOrNull())
-  }
-
-
-    @Test
-  fun addUserButtonFailsIfUserAlreadyInTravel() {
-    val travelContainer = createContainer()
-    listTravelViewModel.selectTravel(travelContainer)
-    composeTestRule.setContent {
-      EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
-    }
-    composeTestRule.onNodeWithTag("importEmailFab").performClick()
-    composeTestRule.onNodeWithTag("addUserFab").performClick()
-
-    // perform add user
-    // Check that the dialog is displayed
-    composeTestRule.onNodeWithTag("roleDialogColumn").assertIsDisplayed()
-    // Check that the title text is displayed and correct
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertTextEquals("Add User by Email")
-    // Check that the OutlinedTextField is displayed and has the correct default value
-    composeTestRule.onNodeWithTag("addUserEmailField").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserEmailField").assertTextContains("")
-
-    val randomEmail = "random.email@example.org"
-    inputText("addUserEmailField", "", randomEmail)
-
-    doAnswer { invocation ->
-          val email = invocation.getArgument<String>(0)
-          val onSuccess = invocation.getArgument<(fsUid?) -> Unit>(1)
-          val customUserInfo =
-              Profile(
-                  fsUid = travelContainer.listParticipant[1],
-                  name = "Custom User",
-                  userTravelList = listOf("00000000000000000000"),
-                  email = email,
-                  username = "username",
-                  friends = null)
-          // Call the onSuccess callback with the custom UserInfo
-          onSuccess(customUserInfo.fsUid)
-        }
-        .`when`(profileRepository)
-        .getFsUidByEmail(any(), any(), any())
-
-    doAnswer { "abcdefghijklmnopqrst" }.`when`(notificationRepository).getNewUid()
-
-    // Mock the repository.updateTravel method to do nothing
-    doNothing().`when`(travelRepository).updateTravel(any(), any(), any())
-    composeTestRule.onNodeWithTag("addUserButton").performClick()
-    verify(profileRepository).getFsUidByEmail(anyOrNull(), anyOrNull(), anyOrNull())
-    verify(notificationRepository, never()).addNotification(anyOrNull())
-  }
-
-  @Test
-  fun addUserButtonWorksIfValidUidAndUserNotInTravel() {
-    val travelContainer = createContainer()
-    listTravelViewModel.selectTravel(travelContainer)
-    composeTestRule.setContent {
-      EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
-    }
-    composeTestRule.onNodeWithTag("importEmailFab").performClick()
-    composeTestRule.onNodeWithTag("addUserFab").performClick()
-
-    // perform add user
-    // Check that the dialog is displayed
-    composeTestRule.onNodeWithTag("roleDialogColumn").assertIsDisplayed()
-    // Check that the title text is displayed and correct
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertTextEquals("Add User by Email")
-    // Check that the OutlinedTextField is displayed and has the correct default value
-    composeTestRule.onNodeWithTag("addUserEmailField").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserEmailField").assertTextContains("")
-
-    val randomEmail = "random.email@example.org"
-    inputText("addUserEmailField", "", randomEmail)
-
-    doAnswer { invocation ->
-          val email = invocation.getArgument<String>(0)
-          val onSuccess = invocation.getArgument<(fsUid?) -> Unit>(1)
-          val customUserInfo =
-              Profile(
-                  fsUid = "qwertzuiopasdfghjklyxcvbnm12",
-                  name = "Custom User",
-                  userTravelList = listOf("00000000000000000000"),
-                  email = email,
-                  username = "username",
-                  friends = null)
-          // Call the onSuccess callback with the custom UserInfo
-          onSuccess(customUserInfo.fsUid)
-        }
-        .`when`(profileRepository)
-        .getFsUidByEmail(any(), any(), any())
-
-    doAnswer { "abcdefghijklmnopqrst" }.`when`(notificationRepository).getNewUid()
-    // Mock the repository.updateTravel method to do nothing
-    doNothing().`when`(travelRepository).updateTravel(any(), any(), any())
-    composeTestRule.onNodeWithTag("addUserButton").performClick()
-    verify(profileRepository).getFsUidByEmail(anyOrNull(), anyOrNull(), anyOrNull())
-
-    verify(notificationRepository).addNotification(anyOrNull())
-  }
-
-  @Test
   fun pressALotOfButtons() {
     val travelContainer = createContainer()
     listTravelViewModel.selectTravel(travelContainer)
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
+    composeTestRule.onNodeWithTag("plusButton").performClick()
+    composeTestRule.onNodeWithTag("manageParticipantsButton").performClick()
+    composeTestRule.onNodeWithTag("goBackButton").performClick()
+
+    composeTestRule.onNodeWithTag("plusButton").performClick()
     composeTestRule.onNodeWithTag("importEmailFab").performClick()
-    composeTestRule.onNodeWithTag("addUserFab").performClick()
-
-    // perform add user
-    // Check that the dialog is displayed
-    composeTestRule.onNodeWithTag("roleDialogColumn").assertIsDisplayed()
-    // Check that the title text is displayed and correct
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertIsDisplayed()
-    composeTestRule.onNodeWithTag("addUserDialogTitle").assertTextEquals("Add User by Email")
-    // Check that the OutlinedTextField is displayed and has the correct default value
-    composeTestRule.onNodeWithTag("addUserEmailField").assertIsDisplayed()
-
-    val randomEmail = "random.email@example.org"
-    inputText("addUserEmailField", "", randomEmail)
-    // Now this is an invalid user that doesn't exist
-    doAnswer { invocation ->
-          val onFailure = invocation.getArgument<(Exception) -> Unit>(2)
-          onFailure(Exception("User not found"))
-        }
-        .`when`(travelRepository)
-        .checkParticipantExists(any(), any(), any())
-    composeTestRule.onNodeWithTag("addUserButton").performClick()
-
-    // Now this is a valid user that had serialisation problems
-    inputText("addUserEmailField", randomEmail, "newuser.email@example.org")
-    doAnswer { invocation ->
-          val onSuccess = invocation.getArgument<(Profile?) -> Unit>(1)
-          // Call the onSuccess callback with null
-          onSuccess(null)
-        }
-        .`when`(travelRepository)
-        .checkParticipantExists(any(), any(), any())
-    // Mock the repository.updateTravel method to do nothing
-    doNothing().`when`(travelRepository).updateTravel(any(), any(), any())
-    composeTestRule.onNodeWithTag("addUserButton").performClick()
-
-    // Now this is a valid user that does exist
-    doAnswer { invocation ->
-          val email = invocation.getArgument<String>(0)
-          val onSuccess = invocation.getArgument<(Profile?) -> Unit>(1)
-          val customUserInfo =
-              Profile(
-                  fsUid = "abcdefghijklmnopqrstuvwxyz12",
-                  name = "Custom User",
-                  userTravelList = listOf("00000000000000000000"),
-                  email = email,
-                  username = "username",
-                  friends = null)
-          // Call the onSuccess callback with the custom UserInfo
-          onSuccess(customUserInfo)
-        }
-        .`when`(travelRepository)
-        .checkParticipantExists(any(), any(), any())
-    // Mock the repository.updateTravel method to do nothing
-    doNothing().`when`(travelRepository).updateTravel(any(), any(), any())
-    composeTestRule.onNodeWithTag("addUserButton").performClick()
+    composeTestRule.onNodeWithTag("goBackButton").performClick()
 
     // perform deletion of travel
     composeTestRule.onNodeWithTag("travelDeleteButton").performClick()
@@ -459,7 +213,11 @@ class EditTravelSettingsScreenTest {
     listTravelViewModel.selectTravel(travelContainer)
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
 
     inputText("inputTravelStartTime", "14/02/2009", "14/02/2009")
@@ -478,7 +236,11 @@ class EditTravelSettingsScreenTest {
     listTravelViewModel.selectTravel(travelContainer)
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
     // check bad dates
     inputText("inputTravelStartTime", "14/02/2009", "14/02/2009")
@@ -488,19 +250,18 @@ class EditTravelSettingsScreenTest {
     composeTestRule.onNodeWithTag("travelSaveButton").performClick()
     inputText("inputTravelEndTime", "13/02/2009", "18/02/2009")
     // check bad locations
-    inputText("inputTravelLatitude", travelContainer.location.latitude.toString(), "gyat")
-    composeTestRule.onNodeWithTag("travelSaveButton").performClick()
 
-    inputText("inputTravelLatitude", "gyat", travelContainer.location.latitude.toString())
-    inputText("inputTravelLongitude", travelContainer.location.longitude.toString(), "-200")
-    composeTestRule.onNodeWithTag("travelSaveButton").performClick()
   }
 
   @Test
   fun backButtonNavigatesCorrectly() {
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
 
     composeTestRule.onNodeWithTag("goBackButton").performClick()
@@ -512,7 +273,11 @@ class EditTravelSettingsScreenTest {
   fun saveButtonPressed() {
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
 
     composeTestRule.onNodeWithTag("travelSaveButton").isDisplayed()
@@ -526,7 +291,11 @@ class EditTravelSettingsScreenTest {
 
     composeTestRule.setContent {
       EditTravelSettingsScreen(
-          listTravelViewModel, navigationActions, notificationViewModel, profileModelView, locationViewModel)
+          listTravelViewModel,
+          navigationActions,
+          notificationViewModel,
+          profileModelView,
+          locationViewModel)
     }
 
     // The startDate picker button should be displayed
